backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
gunicorn
dj-database-url
psycopg2-binary
djangorestframework-simplejwt
python-dotenv
google-auth
requests
